router: moving upstream to a new file (#10253)
