+++ source/common/router/router.h
@@ -28,6 +28,7 @@
 #include "common/config/well_known_names.h"
 #include "common/http/utility.h"
 #include "common/router/config_impl.h"
+#include "common/router/upstream_request.h"
 #include "common/stats/symbol_table_impl.h"
 #include "common/stream_info/stream_info_impl.h"
 #include "common/upstream/load_balancer_impl.h"
@@ -114,6 +115,16 @@ public:
       return r;
     }
   };
+
+  /**
+   * Returns response_time / timeout, as a  percentage as [0, 100]. Returns 0
+   * if there is no timeout.
+   * @param response_time supplies the response time thus far.
+   * @param timeout supplies  the timeout to get the percentage of.
+   * @return the percentage of timeout [0, 100] for stats use.
+   */
+  static uint64_t percentageOfTimeout(const std::chrono::milliseconds response_time,
+                                      const std::chrono::milliseconds timeout);
 
   /**
    * Set the :scheme header based on whether the underline transport is secure.
@@ -231,6 +242,8 @@ private:
 };
 
 using FilterConfigSharedPtr = std::shared_ptr<FilterConfig>;
+
+class UpstreamRequest;
 
 /**
  * Service routing filter.
@@ -363,142 +376,11 @@ public:
     return cookie_value;
   }
 
-protected:
-  RetryStatePtr retry_state_;
-
 private:
-  struct UpstreamRequest : public Http::ResponseDecoder,
-                           public Http::StreamCallbacks,
-                           public Http::ConnectionPool::Callbacks,
-                           public LinkedObject<UpstreamRequest> {
-    UpstreamRequest(Filter& parent, Http::ConnectionPool::Instance& pool);
-    ~UpstreamRequest() override;
-
-    void encodeHeaders(bool end_stream);
-    void encodeData(Buffer::Instance& data, bool end_stream);
-    void encodeTrailers(const Http::RequestTrailerMap& trailers);
-    void encodeMetadata(Http::MetadataMapPtr&& metadata_map_ptr);
-
-    void resetStream();
-    void setupPerTryTimeout();
-    void onPerTryTimeout();
-    void maybeEndDecode(bool end_stream);
-
-    void onUpstreamHostSelected(Upstream::HostDescriptionConstSharedPtr host) {
-      stream_info_.onUpstreamHostSelected(host);
-      upstream_host_ = host;
-      parent_.callbacks_->streamInfo().onUpstreamHostSelected(host);
-      if (parent_.retry_state_ && host) {
-        parent_.retry_state_->onHostAttempted(host);
-      }
-    }
-
-    // Http::StreamDecoder
-    void decodeData(Buffer::Instance& data, bool end_stream) override;
-    void decodeMetadata(Http::MetadataMapPtr&& metadata_map) override;
-
-    // Http::ResponseDecoder
-    void decode100ContinueHeaders(Http::ResponseHeaderMapPtr&& headers) override;
-    void decodeHeaders(Http::ResponseHeaderMapPtr&& headers, bool end_stream) override;
-    void decodeTrailers(Http::ResponseTrailerMapPtr&& trailers) override;
-
-    // Http::StreamCallbacks
-    void onResetStream(Http::StreamResetReason reason,
-                       absl::string_view transport_failure_reason) override;
-    void onAboveWriteBufferHighWatermark() override { disableDataFromDownstream(); }
-    void onBelowWriteBufferLowWatermark() override { enableDataFromDownstream(); }
-
-    void disableDataFromDownstream() {
-      // If there is only one upstream request, we can be assured that
-      // disabling reads will not slow down other upstream requests. If we've
-      // already seen the full downstream request (downstream_end_stream_) then
-      // disabling reads is a no-op.
-      // This assert condition must be true because
-      // parent_.upstream_requests_.size() can only be greater than 1 in the
-      // case of a per-try-timeout with hedge_on_per_try_timeout enabled, and
-      // the per try timeout timer is started only after downstream_end_stream_
-      // is true.
-      ASSERT(parent_.upstream_requests_.size() == 1 || parent_.downstream_end_stream_);
-      parent_.cluster_->stats().upstream_flow_control_backed_up_total_.inc();
-      parent_.callbacks_->onDecoderFilterAboveWriteBufferHighWatermark();
-    }
-
-    void enableDataFromDownstream() {
-      // If there is only one upstream request, we can be assured that
-      // disabling reads will not overflow any write buffers in other upstream
-      // requests. If we've already seen the full downstream request
-      // (downstream_end_stream_) then enabling reads is a no-op.
-      // This assert condition must be true because
-      // parent_.upstream_requests_.size() can only be greater than 1 in the
-      // case of a per-try-timeout with hedge_on_per_try_timeout enabled, and
-      // the per try timeout timer is started only after downstream_end_stream_
-      // is true.
-      ASSERT(parent_.upstream_requests_.size() == 1 || parent_.downstream_end_stream_);
-      parent_.cluster_->stats().upstream_flow_control_drained_total_.inc();
-      parent_.callbacks_->onDecoderFilterBelowWriteBufferLowWatermark();
-    }
-
-    // Http::ConnectionPool::Callbacks
-    void onPoolFailure(Http::ConnectionPool::PoolFailureReason reason,
-                       absl::string_view transport_failure_reason,
-                       Upstream::HostDescriptionConstSharedPtr host) override;
-    void onPoolReady(Http::RequestEncoder& request_encoder,
-                     Upstream::HostDescriptionConstSharedPtr host,
-                     const StreamInfo::StreamInfo& info) override;
-
-    void setRequestEncoder(Http::RequestEncoder& request_encoder);
-    void clearRequestEncoder();
-
-    struct DownstreamWatermarkManager : public Http::DownstreamWatermarkCallbacks {
-      DownstreamWatermarkManager(UpstreamRequest& parent) : parent_(parent) {}
-
-      // Http::DownstreamWatermarkCallbacks
-      void onBelowWriteBufferLowWatermark() override;
-      void onAboveWriteBufferHighWatermark() override;
-
-      UpstreamRequest& parent_;
-    };
-
-    void readEnable();
-
-    Filter& parent_;
-    Http::ConnectionPool::Instance& conn_pool_;
-    bool grpc_rq_success_deferred_;
-    Event::TimerPtr per_try_timeout_;
-    Http::ConnectionPool::Cancellable* conn_pool_stream_handle_{};
-    Http::RequestEncoder* request_encoder_{};
-    absl::optional<Http::StreamResetReason> deferred_reset_reason_;
-    Buffer::WatermarkBufferPtr buffered_request_body_;
-    Upstream::HostDescriptionConstSharedPtr upstream_host_;
-    DownstreamWatermarkManager downstream_watermark_manager_{*this};
-    Tracing::SpanPtr span_;
-    StreamInfo::StreamInfoImpl stream_info_;
-    StreamInfo::UpstreamTiming upstream_timing_;
-    const MonotonicTime start_time_;
-    // Copies of upstream headers/trailers. These are only set if upstream
-    // access logging is configured.
-    Http::ResponseHeaderMapPtr upstream_headers_;
-    Http::ResponseTrailerMapPtr upstream_trailers_;
-    Http::MetadataMapVector downstream_metadata_map_vector_;
-
-    bool calling_encode_headers_ : 1;
-    bool upstream_canary_ : 1;
-    bool decode_complete_ : 1;
-    bool encode_complete_ : 1;
-    bool encode_trailers_ : 1;
-    bool retried_ : 1;
-    bool awaiting_headers_ : 1;
-    bool outlier_detection_timeout_recorded_ : 1;
-    // Tracks whether we deferred a per try timeout because the downstream request
-    // had not been completed yet.
-    bool create_per_try_timeout_on_request_complete_ : 1;
-
-    // Sentinel to indicate if timeout budget tracking is configured for the cluster,
-    // and if so, if the per-try histogram should record a value.
-    bool record_timeout_budget_ : 1;
-  };
-
   using UpstreamRequestPtr = std::unique_ptr<UpstreamRequest>;
+  friend class UpstreamRequest;
+
+  RetryStatePtr retry_state_;
 
   StreamInfo::ResponseFlag streamResetReasonToResponseFlag(Http::StreamResetReason reset_reason);
