+++ source/common/router/upstream_request.h
@@ -1,0 +1,148 @@
+#pragma once
+
+#include <chrono>
+#include <cstdint>
+#include <functional>
+#include <memory>
+#include <string>
+
+#include "envoy/http/codec.h"
+#include "envoy/http/codes.h"
+#include "envoy/http/filter.h"
+#include "envoy/stats/scope.h"
+
+#include "common/buffer/watermark_buffer.h"
+#include "common/common/cleanup.h"
+#include "common/common/hash.h"
+#include "common/common/hex.h"
+#include "common/common/linked_object.h"
+#include "common/common/logger.h"
+#include "common/config/well_known_names.h"
+#include "common/router/router.h"
+#include "common/stream_info/stream_info_impl.h"
+
+namespace Envoy {
+namespace Router {
+
+class Filter;
+
+class UpstreamRequest : public Logger::Loggable<Logger::Id::router>,
+                        public Http::ResponseDecoder,
+                        public Http::StreamCallbacks,
+                        public Http::ConnectionPool::Callbacks,
+                        public LinkedObject<UpstreamRequest> {
+public:
+  UpstreamRequest(Filter& parent, Http::ConnectionPool::Instance& pool);
+  ~UpstreamRequest() override;
+
+  void encodeHeaders(bool end_stream);
+  void encodeData(Buffer::Instance& data, bool end_stream);
+  void encodeTrailers(const Http::RequestTrailerMap& trailers);
+  void encodeMetadata(Http::MetadataMapPtr&& metadata_map_ptr);
+
+  void resetStream();
+  void setupPerTryTimeout();
+  void onPerTryTimeout();
+  void maybeEndDecode(bool end_stream);
+  void onUpstreamHostSelected(Upstream::HostDescriptionConstSharedPtr host);
+
+  // Http::StreamDecoder
+  void decodeData(Buffer::Instance& data, bool end_stream) override;
+  void decodeMetadata(Http::MetadataMapPtr&& metadata_map) override;
+
+  // Http::ResponseDecoder
+  void decode100ContinueHeaders(Http::ResponseHeaderMapPtr&& headers) override;
+  void decodeHeaders(Http::ResponseHeaderMapPtr&& headers, bool end_stream) override;
+  void decodeTrailers(Http::ResponseTrailerMapPtr&& trailers) override;
+
+  // Http::StreamCallbacks
+  void onResetStream(Http::StreamResetReason reason,
+                     absl::string_view transport_failure_reason) override;
+  void onAboveWriteBufferHighWatermark() override { disableDataFromDownstreamForFlowControl(); }
+  void onBelowWriteBufferLowWatermark() override { enableDataFromDownstreamForFlowControl(); }
+
+  void disableDataFromDownstreamForFlowControl();
+  void enableDataFromDownstreamForFlowControl();
+
+  // Http::ConnectionPool::Callbacks
+  void onPoolFailure(Http::ConnectionPool::PoolFailureReason reason,
+                     absl::string_view transport_failure_reason,
+                     Upstream::HostDescriptionConstSharedPtr host) override;
+  void onPoolReady(Http::RequestEncoder& request_encoder,
+                   Upstream::HostDescriptionConstSharedPtr host,
+                   const StreamInfo::StreamInfo& info) override;
+
+  void setRequestEncoder(Http::RequestEncoder& request_encoder);
+  void clearRequestEncoder();
+
+  struct DownstreamWatermarkManager : public Http::DownstreamWatermarkCallbacks {
+    DownstreamWatermarkManager(UpstreamRequest& parent) : parent_(parent) {}
+
+    // Http::DownstreamWatermarkCallbacks
+    void onBelowWriteBufferLowWatermark() override;
+    void onAboveWriteBufferHighWatermark() override;
+
+    UpstreamRequest& parent_;
+  };
+
+  void readEnable();
+
+  // Getters and setters
+  Upstream::HostDescriptionConstSharedPtr& upstreamHost() { return upstream_host_; }
+  void outlierDetectionTimeoutRecorded(bool recorded) {
+    outlier_detection_timeout_recorded_ = recorded;
+  }
+  bool outlierDetectionTimeoutRecorded() { return outlier_detection_timeout_recorded_; }
+  const StreamInfo::UpstreamTiming& upstreamTiming() { return upstream_timing_; }
+  void retried(bool value) { retried_ = value; }
+  bool retried() { return retried_; }
+  bool grpcRqSuccessDeferred() { return grpc_rq_success_deferred_; }
+  void grpcRqSuccessDeferred(bool deferred) { grpc_rq_success_deferred_ = deferred; }
+  void upstreamCanary(bool value) { upstream_canary_ = value; }
+  bool upstreamCanary() { return upstream_canary_; }
+  bool awaitingHeaders() { return awaiting_headers_; }
+  void recordTimeoutBudget(bool value) { record_timeout_budget_ = value; }
+  bool createPerTryTimeoutOnRequestComplete() {
+    return create_per_try_timeout_on_request_complete_;
+  }
+
+private:
+  Filter& parent_;
+  Http::ConnectionPool::Instance& conn_pool_;
+  bool grpc_rq_success_deferred_;
+  Event::TimerPtr per_try_timeout_;
+  Http::ConnectionPool::Cancellable* conn_pool_stream_handle_{};
+  Http::RequestEncoder* request_encoder_{};
+  absl::optional<Http::StreamResetReason> deferred_reset_reason_;
+  Buffer::WatermarkBufferPtr buffered_request_body_;
+  Upstream::HostDescriptionConstSharedPtr upstream_host_;
+  DownstreamWatermarkManager downstream_watermark_manager_{*this};
+  Tracing::SpanPtr span_;
+  StreamInfo::StreamInfoImpl stream_info_;
+  StreamInfo::UpstreamTiming upstream_timing_;
+  const MonotonicTime start_time_;
+  // Copies of upstream headers/trailers. These are only set if upstream
+  // access logging is configured.
+  Http::ResponseHeaderMapPtr upstream_headers_;
+  Http::ResponseTrailerMapPtr upstream_trailers_;
+  Http::MetadataMapVector downstream_metadata_map_vector_;
+
+  bool calling_encode_headers_ : 1;
+  bool upstream_canary_ : 1;
+  bool decode_complete_ : 1;
+  bool encode_complete_ : 1;
+  bool encode_trailers_ : 1;
+  bool retried_ : 1;
+  bool awaiting_headers_ : 1;
+  bool outlier_detection_timeout_recorded_ : 1;
+  // Tracks whether we deferred a per try timeout because the downstream request
+  // had not been completed yet.
+  bool create_per_try_timeout_on_request_complete_ : 1;
+
+  // Sentinel to indicate if timeout budget tracking is configured for the cluster,
+  // and if so, if the per-try histogram should record a value.
+  bool record_timeout_budget_ : 1;
+};
+
+} // namespace Router
+} // namespace Envoy
